requirements.txt
pyslack
slackclient
Adafruit-DHT